Add Student page refactored
