+++ src/pages/CreateClass/index.js
@@ -1,158 +1,48 @@
-import React, { useState, useEffect, useRef } from "react";
-import { useHistory } from "react-router-dom";
-import CreateClassComponent from "../../components/Class";
-import ClassesList from "../../components/Class/ClassList";
+import React, { useState, useEffect } from "react";
 import { useSelector } from "react-redux";
-import "../../components/Class/ClassList/styles.scss";
-import axios from "axios";
-import { METHODS } from "../../services/api";
-import "./styles.scss";
-import useStyles from "./styles";
-import {
-  Container,
-  Button,
-  Modal,
-  useMediaQuery,
-  Box,
-  Stack,
-  Typography,
-  Dialog,
-  DialogTitle,
-  DialogActions,
-} from "@mui/material";
+import { Container, Button, Box, useMediaQuery, Modal } from "@mui/material";
 import { breakpoints } from "../../theme/constant";
-import ClassForm from "../../components/Class/ClassForm";
-
-import SuccessModel from "../../components/Class/SuccessModel";
 import NewVolunteerCard from "../../components/Class/NewVolunteerCard";
+import SuccessModel from "../../components/Class/SuccessModel";
+import ClassesList from "../../components/Class/ClassList";
+import ClassFormModal from "./ClassFormModal";
 
-function ToggleClassFormModal() {
+function ClassManagementPage() {
   const [showModal, setShowModal] = useState(false);
   const [formType, setFormType] = useState("batch");
   const [classToEdit, setClassToEdit] = useState({});
   const [indicator, setIndicator] = useState(false);
-  const [showConsentModal, setShowConsentModal] = useState(false);
-  const { data = [] } = useSelector(({ Class }) => Class.allClasses);
-  const user = useSelector(({ User }) => User);
-  const [open, setOpen] = React.useState(false);
-  const handleOpen = () => setOpen(true);
   const [openSuccessfullModal, setOpenSuccessfullModal] = useState(false);
-  const [isEditMode, setIsEditMode] = React.useState(false);
-  const classes = useStyles();
 
   const isActive = useMediaQuery("(max-width:" + breakpoints.values.sm + "px)");
   const isActiveIpad = useMediaQuery("(max-width:1300px)");
 
-  const rolesList = user.data.user.rolesList;
-  const canSpecifyFacilitator = rolesList.indexOf("admin") > -1;
+  const { data = [] } = useSelector(({ Class }) => Class.allClasses);
+  const user = useSelector(({ User }) => User);
 
-  const [calenderConsent, setCalenderConsent] = useState(true);
-  const [authUrl, setAuthUrl] = useState("");
+  const canSpecifyFacilitator = user.data.user.rolesList.includes("admin");
 
-  const url = window.location.href;
-
-  const toggleModalOpen = () => {
-    // setFormType();
-    setClassToEdit({});
-    // setShowModal(!showModal);
-    CalenderConsent();
-  };
-
-  //here can check
-  const editClass = (classId, indicator) => {
-    setClassToEdit(data.find((classData) => classData.id === classId));
-    setIsEditMode(true);
-    setShowModal(true);
-    setIndicator(indicator);
-  };
-
-  const CalenderConsent = () => {
-    axios({
-      method: METHODS.GET,
-      url: `${process.env.REACT_APP_MERAKI_URL}/users/calendar/tokens`,
-      headers: {
-        accept: "application/json",
-        Authorization: user.data.token,
-      },
-    }).then((res) => {
-      if (res.data.success) {
-        setCalenderConsent(true);
-        setShowModal(true);
-      } else {
-        setCalenderConsent(false);
-        setShowConsentModal(true);
-        setShowModal(false);
-      }
-    });
-  };
-
-  const handleClose = () => {
-    setShowConsentModal(false);
-  };
-
-  const codeGenerate = async () => {
-    return axios({
-      method: METHODS.GET,
-      url: `${process.env.REACT_APP_MERAKI_URL}/users/calendar/generateAuthURL`,
-      headers: {
-        accept: "application/json",
-        Authorization: user.data.token,
-      },
-    }).then((res) => {
-      setAuthUrl(res.data.url);
-    });
-  };
-
-  // console.log("authUrl", authUrl);
-
-  const calledOnce = useRef(false);
-  const history = useHistory();
-
-  useEffect(() => {
-    let code;
-    let payload;
-    let user_id;
-    let user_email;
-    if (url.includes("code")) {
-      const decodedUri = decodeURIComponent(url);
-      user_id = decodedUri.split("=")[2].split("+")[0];
-      user_email = decodedUri.split("=")[3].split("&")[0];
-      code = url.split("code=")[1].split("scope")[0];
-      payload = {
-        ...payload,
-        user_id: parseInt(user_id, 10),
-        user_email: user_email,
-      };
-      calledOnce.current = true;
-    }
-    if (calledOnce.current) {
-      return axios({
-        method: METHODS.PUT,
-        url: `${process.env.REACT_APP_MERAKI_URL}/users/calendar/tokens`,
-
-        headers: {
-          accept: "application/json",
-          Authorization: user.data.token,
-          code: code,
-        },
-        data: payload,
-      }).then((res) => {
-        if (res.data.success) {
-          setShowModal(true);
-          history.push("/class");
-        }
-      });
-    }
-  }, [calledOnce]);
   const [newVolunteer, setNewVolunteer] = useState(false);
+
   useEffect(() => {
     const newVol = localStorage.getItem("isNewVolunteer");
-    if (newVol == "true" && newVol != null) {
+    if (newVol === "true" && newVol != null) {
       setNewVolunteer(true);
     } else {
       setNewVolunteer(false);
     }
-  }, [newVolunteer]);
+  }, []);
+
+  const toggleModalOpen = () => {
+    setShowModal(true);
+  };
+
+  const editClass = (classId, indicator) => {
+    setClassToEdit(data.find((classData) => classData.id === classId));
+    setShowModal(true);
+    setIndicator(indicator);
+  };
+
   return (
     <Container maxWidth="lg" sx={{ mt: "40px", width: "90%" }}>
       {canSpecifyFacilitator && (
@@ -191,77 +81,16 @@ function ToggleClassFormModal() {
           </Box>
         </span>
       )}
-      <ClassesList
-        editClass={editClass}
-        isShow={showModal}
-        // setIsEditMode={setIsEditMode}
+      <ClassesList editClass={editClass} />
+      <ClassFormModal
+        showModal={showModal}
+        setShowModal={setShowModal}
+        isEditMode={true}
+        indicator={indicator}
+        classToEdit={classToEdit}
+        formType={formType}
+        setOpenSuccessfullModal={setOpenSuccessfullModal}
       />
-      {showModal && calenderConsent ? (
-        <Modal
-          open={showModal}
-          aria-labelledby="modal-modal-title"
-          aria-describedby="modal-modal-description"
-          style={{ overflow: "scroll" }}
-        >
-          <ClassForm
-            isEditMode={isEditMode}
-            indicator={indicator}
-            classToEdit={classToEdit}
-            formType={formType}
-            setShowModal={setShowModal}
-            setOpenSuccessfullModal={setOpenSuccessfullModal}
-          />
-          {/* <CreateClassComponent
-            classToEdit={classToEdit}
-            indicator={indicator}
-            toggleModalOpen={toggleModalOpen}
-          /> */}
-        </Modal>
-      ) : (
-        showConsentModal && (
-          <Dialog
-            open={showConsentModal}
-            aria-labelledby="alert-dialog-title"
-            aria-describedby="alert-dialog-description"
-            PaperProps={{
-              style: {
-                minWidth: "45%",
-                borderRadius: 8,
-              },
-            }}
-          >
-            <DialogTitle>
-              <Typography variant="h6" align="center">
-                Meraki needs access to your calendar to create classes. <br />
-                Do you want to go ahead?
-              </Typography>
-            </DialogTitle>
-            <Stack alignItems="center">
-              <DialogActions>
-                <Box sx={{ display: "flex", mb: 2 }}>
-                  <Button
-                    onClick={codeGenerate}
-                    color="error"
-                    variant="contained"
-                    sx={{ mr: "15px", width: "100px" }}
-                  >
-                    Yes
-                  </Button>
-                  <Button
-                    onClick={handleClose}
-                    color="grey"
-                    variant="contained"
-                    sx={{ width: "100px" }}
-                  >
-                    No
-                  </Button>
-                </Box>
-              </DialogActions>
-            </Stack>
-          </Dialog>
-        )
-      )}
-
       {openSuccessfullModal && (
         <Modal
           open={openSuccessfullModal}
@@ -272,10 +101,8 @@ function ToggleClassFormModal() {
           <SuccessModel />
         </Modal>
       )}
-
-      {authUrl && (window.location.href = authUrl)}
     </Container>
   );
 }
 
-export default ToggleClassFormModal;
+export default ClassManagementPage;
