--- src/pages/GsocIdea/styles.js
@@ -1,15 +1,0 @@
-import { makeStyles } from "@mui/styles";
-
-const useStyles = makeStyles((theme) => ({
-  listItem: {
-    fontSize: "1.125rem",
-    fontFamily: "Noto Sans",
-    fontWeight: "400",
-    lineHeight: "1.5",
-  },
-  headingMargin: {
-    marginTop: "32px",
-  },
-}));
-
-export default useStyles;
