+++ src/routing/index.js
@@ -34,7 +34,6 @@ import ReturningUserPage from "../components/ReturningUser/ReturningUserPage";
 import VolunteerAutomation from "../components/VolunteerAutomation";
 import VolunteerForm from "../components/VolunteerAutomation/Stepper";
 // import ClassFormModel from "../components/BatchClassComponents/ClassFormModel";
-import GsocIdea from "../pages/GsocIdea/GsocIdea";
 
 const Routing = () => {
   return (
@@ -83,8 +82,7 @@ const Routing = () => {
         path={PATHS.VOLUNTEER_AUTOMATION}
         component={VolunteerAutomation}
       />
-      <Route exact path={PATHS.GSOC_IDEA} component={GsocIdea} />
-      
+
       {/* Private routes */}
 
       <PrivateRoute
